+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2022-11-10-1
+
+Changes:
+
+- Fix select language `force=false`
+
 ## v2022-11-09-1
 
 Changes: